+++ lib/wavefront-cli/controller.rb
@@ -1,7 +1,7 @@
 # For development against a local checkout of the SDK, uncomment
 # this definition
 #
-# DEVELOPMENT = true
+ DEVELOPMENT = true
 
 if defined?(DEVELOPMENT)
   dir = Pathname.new(__FILE__).dirname.realpath.parent.parent.parent