+++ lib/wavefront-cli/display/dashboard.rb
@@ -32,5 +32,22 @@ module WavefrontDisplay
     def do_unfav
       puts "Removed #{options[:'<id>']} from favourites."
     end
+
+    def do_acls
+      data.each do |dash|
+        display_acl('view and modify', dash[:modifyAcl])
+        display_acl('view', dash[:viewAcl])
+      end
+    end
+
+    def display_acl(title, acl_data)
+      puts title
+
+      if acl_data.empty?
+        puts '  <none>'
+      else
+        acl_data.each { |e| puts format('  %<name>s (%<id>s)', e) }
+      end
+    end
   end
 end